+++ mmlmtools/adapter/igpt.py
@@ -16,6 +16,6 @@ def convert_mmtools_for_igpt(models):
 
 # global_dict = sys.modules['iGPT.controllers.ConversationBot'].__dict__
 # for k, v in tools.__dict__.items():
-#     if inspect.isclass(v) and issubclass(v, BaseTool):
+#     if inspect.isclass(v) and issubclass(v, BaseToolv1):
 #         global_dict[k] = v
 import_all_tools_to('iGPT.controllers.ConversationBot')
