+++ mmlmtools/tools/edge_detection.py
@@ -5,10 +5,10 @@ from PIL import Image
 
 from mmlmtools.toolmeta import ToolMeta
 from ..utils.utils import get_new_image_name
-from .base_tool import BaseTool
+from .base_tool_v1 import BaseToolv1
 
 
-class Image2CannyTool(BaseTool):
+class Image2CannyTool(BaseToolv1):
     DEFAULT_TOOLMETA = dict(
         name='Edge Detection On Image',
         model=None,
