+++ mmlmtools/tools/stylization.py
@@ -7,10 +7,10 @@ from PIL import Image
 
 from mmlmtools.toolmeta import ToolMeta
 from ..utils.utils import get_new_image_name
-from .base_tool import BaseTool
+from .base_tool_v1 import BaseToolv1
 
 
-class InstructPix2PixTool(BaseTool):
+class InstructPix2PixTool(BaseToolv1):
     DEFAULT_TOOLMETA = dict(
         name='Instruct Image Using Text',
         model={'model_name': 'timbrooks/instruct-pix2pix'},
